Fix #393 The description of shows does not appear.
This fixes a problem with type comparison in PHP 7.

In line 279 the variable $atts['view'] is initialized with 0 as default.

In line 753 of `includes/shortcodes.php` there exists a type comparison:

` 'grid' == $atts['view']`

Because PHP 7 considers the integer 0 to be equal to a string, the show description was not shown in the shortcode. I fixed this by changing the default value from 0 to 'none'.